+++ tests/e2e/utils/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `utils.waitForTimeout( delay )` pause processing for `delay` milliseconds
 - `AdminEdit` class with utility functions for the respective edit screens
+- Update `shopper.addToCartFromShopPage()` and `.removeFromCart()` to accept product Id or Title
 
 # 0.1.6